+++ core/src/scheduler/selector/FaissIVFPQPass.cpp
@@ -27,7 +27,7 @@ void
 FaissIVFPQPass::Init() {
 #ifdef MILVUS_GPU_VERSION
     server::Config& config = server::Config::GetInstance();
-    Status s = config.GetEngineConfigGpuSearchThreshold(threshold_);
+    Status s = config.GetGpuResourceConfigGpuSearchThreshold(threshold_);
     if (!s.ok()) {
         threshold_ = std::numeric_limits<int32_t>::max();
     }