+++ core/src/server/init/InstanceLockCheck.cpp
@@ -10,12 +10,15 @@
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
 #include "server/init/InstanceLockCheck.h"
-#include "utils/Log.h"
 
 #include <errno.h>
 #include <fcntl.h>
 #include <string.h>
 #include <unistd.h>
+
+#include <fiu-local.h>
+
+#include "utils/Log.h"
 
 namespace milvus {
 namespace server {
@@ -24,6 +27,7 @@ Status
 InstanceLockCheck::Check(const std::string& path) {
     std::string lock_path = path + "/lock";
     auto fd = open(lock_path.c_str(), O_RDWR | O_CREAT | O_NOFOLLOW, 0640);
+    fiu_do_on("InstanceLockCheck.Check.fd", fd = -1);
     if (fd < 0) {
         std::string msg;
         if (errno == EROFS) {
@@ -41,7 +45,9 @@ InstanceLockCheck::Check(const std::string& path) {
     fl.l_whence = SEEK_SET;
     fl.l_start = 0;
     fl.l_len = 0;
-    if (fcntl(fd, F_SETLK, &fl) == -1) {
+    auto fcl = fcntl(fd, F_SETLK, &fl);
+    fiu_do_on("InstanceLockCheck.Check.fcntl", fcl = -1);
+    if (fcl == -1) {
         std::string msg = "Can't lock file: " + lock_path + ", due to ";
         if (errno == EACCES || errno == EAGAIN) {
             msg += "permission denied. ";