+++ CHANGELOG.md
@@ -1,23 +1,28 @@
 # Changelog
 
 Please mark all change in change log and use the issue from GitHub
+
 # Milvus 0.10.0 (TBD)
 
 ## Bug
 -   \#2367 Fix inconsistent reading and writing when using mishards
 -   \#2368 Make read node detect delete behavior
 -   \#2394 Drop collection timeout if too many partitions created on collection
+-   \#2549 Launch server fail using demo config
 
 ## Feature
 -   \#2363 Update branch version
+-   \#2510 Upgrade Milvus config
 
 ## Improvement
 -   \#2381 Upgrade FAISS to 1.6.3
 -   \#2429 Fix Milvus 0.9.1 performance degrade issue
 -   \#2441 Improve Knowhere code coverage
--   \#2466 optimize k-selection implemention of faiss gpu version
+-   \#2466 Optimize k-selection implemention of faiss gpu version
 -   \#2489 Add exception throw on mysql meta error
 -   \#2495 Add creating lock file failure reason.
+-   \#2516 Improve unit test coverage
+-   \#2548 Upgrade mishards for milvus v0.10.0
 
 ## Task